fix "jumping to first slide" issue
this is a hacky fix to some hack-y code, but i've removed the reset to index 0 every time the slideshow is started, which means if a room is opened up anew it will have the slide set to whatever it was before. a "set to first slide" button has been added to make it easier to change.